重要更新：新增换源功能、横向阅读功能、搜索排序功能

+++ app/src/main/java/com/Z/NovelReader/NovelSourceManageActivity.java
@@ -2,10 +2,11 @@ package com.Z.NovelReader;
 
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
+import androidx.appcompat.app.AlertDialog;
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.core.content.ContextCompat;
 import androidx.lifecycle.Observer;
-import androidx.lifecycle.ViewModelProviders;
+import androidx.lifecycle.ViewModelProvider;
 
 import android.app.Activity;
 import android.content.Context;
@@ -30,16 +31,25 @@ import android.widget.TextView;
 import android.widget.Toast;
 
 import com.Z.NovelReader.Adapters.NovelSourceAdapter;
+import com.Z.NovelReader.Basic.BasicCounterHandler;
+import com.Z.NovelReader.Basic.BasicHandler;
+import com.Z.NovelReader.NovelRoom.NovelDBUpdater;
+import com.Z.NovelReader.NovelRoom.Novels;
 import com.Z.NovelReader.NovelSourceRoom.NovelSourceDBTools;
 import com.Z.NovelReader.NovelSourceRoom.NovelSourceViewModel;
+import com.Z.NovelReader.Threads.CheckRespondThread;
 import com.Z.NovelReader.Threads.NovelSourceGetterThread;
 import com.Z.NovelReader.Utils.FileIOUtils;
+import com.Z.NovelReader.Utils.QRcodeUtils.Constant;
 import com.Z.NovelReader.Utils.ScreenUtils;
 import com.Z.NovelReader.Utils.StatusBarUtil;
 import com.Z.NovelReader.Utils.StorageUtils;
 import com.Z.NovelReader.Objects.beans.NovelRequire;
 import com.Z.NovelReader.views.KeyboardPopupWindow;
-import com.Z.NovelReader.views.WaitDialog;
+import com.Z.NovelReader.views.Dialog.WaitDialog;
+import com.Z.NovelReader.views.NovelSourceMenu;
+import com.Z.NovelReader.zxing.activity.CaptureActivity;
+import com.Z.NovelReader.zxing.camera.CameraManager;
 import com.google.gson.JsonSyntaxException;
 import com.z.fileselectorlib.FileSelectorSettings;
 import com.z.fileselectorlib.Objects.FileInfo;
@@ -47,6 +57,7 @@ import com.z.fileselectorlib.Objects.FileInfo;
 import java.io.File;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Locale;
 
 public class NovelSourceManageActivity extends AppCompatActivity {
 
@@ -54,18 +65,22 @@ public class NovelSourceManageActivity extends AppCompatActivity {
     private Window window;
     private ListView lv_novelSourceList;
     private ImageButton im_manage;
-    private PopupWindow manageOptions;
+    //private PopupWindow manageOptions;
+    private NovelSourceMenu source_manage_menu;
     private WaitDialog waitDialog;
     //data sources
     private NovelSourceAdapter adapter;
     private NovelSourceDBTools sourceDBTools;//书源数据库DAO
+    private NovelDBUpdater novelDBUpdater;//书籍数据库DAO
     private NovelSourceViewModel novelSourceViewModel;//书源数据维护
     private List<NovelRequire> novelRequireList;//书源列表
     //thread & handler
-    private NovelSourceGetterThread.NSGetterHandler handler;
+    //private NovelSourceGetterThread.NSGetterHandler handler;
+    private BasicHandler<NovelRequire> handler;
     //basic paras
     private Context context;
     private Activity activity;
+    private boolean add_source_flag = false;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -81,11 +96,13 @@ public class NovelSourceManageActivity extends AppCompatActivity {
         //界面初始化
         initStatusBar();
         //init handler
-        handler = new NovelSourceGetterThread.NSGetterHandler(new NovelSourceGetterThread.NSGetterListener() {
+        handler = new BasicHandler<>(new BasicHandler.BasicHandlerListener<NovelRequire>() {
             @Override
-            public void onSuccess() {
+            public void onSuccess(NovelRequire result) {
                 if (waitDialog!=null)waitDialog.dismiss();
                 Toast.makeText(context, "书源添加完成", Toast.LENGTH_SHORT).show();
+                CheckRespondThread thread = new CheckRespondThread(result,sourceDBTools);
+                thread.start();
             }
 
             @Override
@@ -102,13 +119,42 @@ public class NovelSourceManageActivity extends AppCompatActivity {
                 }
             }
         });
+//        handler = new NovelSourceGetterThread.NSGetterHandler(new NovelSourceGetterThread.NSGetterListener() {
+//            @Override
+//            public void onSuccess() {
+//                if (waitDialog!=null)waitDialog.dismiss();
+//                Toast.makeText(context, "书源添加完成", Toast.LENGTH_SHORT).show();
+//            }
+//
+//            @Override
+//            public void onError(int error_code) {
+//                if (waitDialog!=null)waitDialog.dismiss();
+//                switch(error_code){
+//                    case NovelSourceGetterThread.NO_INTERNET:
+//                        Toast.makeText(context, "无网络", Toast.LENGTH_SHORT).show();
+//                        break;
+//                    case NovelSourceGetterThread.SOURCE_TYPE_NOT_JSON:
+//                        Toast.makeText(context, "书源链接无效:不是json", Toast.LENGTH_SHORT).show();
+//                        break;
+//                    default:
+//                }
+//            }
+//        });
+
         //书源数据库DAO初始化
         sourceDBTools = new NovelSourceDBTools(this);
+        novelDBUpdater = new NovelDBUpdater(this);
         //初始化view model，启动observer
-        novelSourceViewModel = ViewModelProviders.of(this).get(NovelSourceViewModel.class);
+        novelSourceViewModel = new ViewModelProvider(this,ViewModelProvider.AndroidViewModelFactory.getInstance(getApplication())).get(NovelSourceViewModel.class);
         novelSourceViewModel.getNovelSourceLiveData().observe(this, new Observer<List<NovelRequire>>() {
             @Override
             public void onChanged(List<NovelRequire> novelRequires) {
+                if (add_source_flag){
+                    String info = String.format(Locale.CHINA,"新增：%d个书源",
+                            novelRequires.size()-novelRequireList.size());
+                    Toast.makeText(context, info, Toast.LENGTH_SHORT).show();
+                    add_source_flag = false;
+                }
                 novelRequireList=novelRequires;
                 adapter.setNovelRequireList(novelRequireList);
                 adapter.notifyDataSetChanged();
@@ -130,7 +176,27 @@ public class NovelSourceManageActivity extends AppCompatActivity {
 
             @Override
             public void onSourceDelete(NovelRequire currentSource) {
-                sourceDBTools.DeleteNovelSources(currentSource);
+                try {
+                    List<Novels> novels = novelDBUpdater.QueryNovelsBySource(currentSource.getId());
+                    if (novels.size()!=0) {
+                        AlertDialog.Builder builder = new AlertDialog.Builder(activity);
+                        builder.setTitle("删除书源")
+                                .setMessage(String.format(Locale.CHINA, "书架中将有 %d本书被一并删除", novels.size()))
+                                .setPositiveButton("确定", (dialogInterface, i) -> {
+                                    novelDBUpdater.deleteNovels(novels.toArray(new Novels[0]));
+                                    sourceDBTools.DeleteNovelSources(currentSource);
+                                })
+                                .setNegativeButton("取消", ((dialogInterface, i) -> {
+                                    Log.d("novel source manage", "cancel delete");
+                                }))
+                                .setCancelable(false).show();
+                    }else{
+                        sourceDBTools.DeleteNovelSources(currentSource);
+                    }
+                } catch (InterruptedException e) {
+                    e.printStackTrace();
+                }
+
             }
         });
         //list 显示数据
@@ -155,15 +221,21 @@ public class NovelSourceManageActivity extends AppCompatActivity {
         super.onRequestPermissionsResult(requestCode, permissions, grantResults);
         if(grantResults[0]== PackageManager.PERMISSION_GRANTED){
             //权限重新允许
-            showFileSelector();
-            manageOptions.dismiss();
+            if (requestCode == StorageUtils.STORAGE_RQS_CODE)
+                showFileSelector();
+            if (requestCode == Constant.REQ_PERM_CAMERA){
+                Intent intent = new Intent(NovelSourceManageActivity.this, CaptureActivity.class);
+                startActivityForResult(intent, Constant.REQ_QR_CODE);
+            }
+            source_manage_menu.dismiss();
         }
     }
 
     @Override
     protected void onActivityResult(int requestCode, int resultCode, @Nullable Intent data) {
         super.onActivityResult(requestCode, resultCode, data);
-        if (requestCode==FileSelectorSettings.REQUEST_CODE && resultCode==FileSelectorSettings.BACK_WITH_SELECTIONS){
+        //本地文件导入回调
+        if (requestCode==FileSelectorSettings.FILELIST_REQUEST_CODE && resultCode==FileSelectorSettings.BACK_WITH_SELECTIONS){
             assert data != null;
             Bundle bundle=data.getExtras();
             assert bundle != null;
@@ -172,74 +244,110 @@ public class NovelSourceManageActivity extends AppCompatActivity {
             if (FilePathSelected!=null && FilePathSelected.size()!=0) {
                 String path=FilePathSelected.get(0);
                 File local_source = new File(path);
-                if (local_source.exists()){
-                    try {
-                        String sourceJSON = FileIOUtils.read_line("", local_source);
-                        NovelRequire[] novelRequire = NovelRequire.getNovelRequireBeans(sourceJSON);
-                        NovelSourceDBTools sourceDBTools=new NovelSourceDBTools(context);
-                        sourceDBTools.InsertNovelSources(novelRequire);
-                        Toast.makeText(this, "新增：" +novelRequire.length+ "个书源", Toast.LENGTH_SHORT).show();
-                    }catch (JsonSyntaxException e){
-                        e.printStackTrace();
-                        Toast.makeText(context, "文件不符合书源格式", Toast.LENGTH_SHORT).show();
-                    }
+                addNovelRequire(local_source);
+            }
+        }
+        //二维码导入回调
+        if (requestCode == Constant.REQ_QR_CODE && resultCode == RESULT_OK) {
+            Bundle bundle = data.getExtras();
+            String scanResult = bundle.getString(Constant.INTENT_EXTRA_KEY_QR_SCAN);
+            //将扫描出的信息显示出来
+            Log.d("qr code",""+scanResult);
+            if (!scanResult.contains("http")){
+                Toast.makeText(context, "未读取到链接", Toast.LENGTH_SHORT).show();
+                return;
+            }
+            NovelSourceGetterThread sourceGetterThread=new NovelSourceGetterThread(context,scanResult);
+            sourceGetterThread.setHandler(handler);
+            sourceGetterThread.start();
+        }
+    }
+
+    private void addNovelRequire(File local_source) {
+        if (local_source.exists()){
+            try {
+                String sourceJSON = FileIOUtils.read_line(local_source.getPath());
+                NovelRequire[] novelRequire = NovelRequire.getNovelRequireBeans(sourceJSON);
+                NovelSourceDBTools sourceDBTools=new NovelSourceDBTools(context);
+                sourceDBTools.InsertNovelSources(novelRequire);
+                //check source
+                for (NovelRequire rule : novelRequire) {
+                    CheckRespondThread thread = new CheckRespondThread(rule,sourceDBTools);
+                    thread.start();
                 }
+                add_source_flag = true;
+            }catch (JsonSyntaxException e){
+                e.printStackTrace();
+                Toast.makeText(context, "文件不符合书源格式", Toast.LENGTH_SHORT).show();
             }
         }
     }
 
     private void showManageOptions() {
-        //加载布局
-        RelativeLayout layout = (RelativeLayout) LayoutInflater.from(activity).inflate(R.layout.novel_source_manage_popup, null);
-        //找到布局的控件
-        final LinearLayout ll_internet_import = layout.findViewById(R.id.import_from_internet);
-        final LinearLayout ll_local_import = layout.findViewById(R.id.import_from_local);
-        final LinearLayout ll_delete_select = layout.findViewById(R.id.delete_select);
-        final TextView delete_select_title = ll_delete_select.findViewById(R.id.delete_select_title);
-        //设置属性
-        if (adapter.isDelete_mode()){
-            delete_select_title.setText("退出删除");
-        }else {
-            delete_select_title.setText("删除模式");
-        }
-        // 实例化popupWindow
-        manageOptions = new PopupWindow(layout, ViewGroup.LayoutParams.WRAP_CONTENT, ViewGroup.LayoutParams.WRAP_CONTENT);
-        //焦点和阴影
-        manageOptions.setFocusable(true);
-        manageOptions.setElevation(8);
-        //设置popupWindow弹出窗体位置
-        layout.measure(View.MeasureSpec.UNSPECIFIED, View.MeasureSpec.UNSPECIFIED);
-        manageOptions.setBackgroundDrawable(ContextCompat.getDrawable(context,R.drawable.popwindow_white));
-        manageOptions.showAsDropDown(im_manage,-layout.getMeasuredWidth() + im_manage.getWidth(),15);
-        //点击事件
-        //网络导入
-        ll_internet_import.setOnClickListener(new View.OnClickListener() {
+        source_manage_menu = new NovelSourceMenu(context,adapter.isDelete_mode());
+        source_manage_menu.setMenuClickListener(new NovelSourceMenu.MenuClickListener() {
             @Override
-            public void onClick(View v) {
+            public void doInternetImport() {
                 showEditTextForURL();
             }
-        });
-        ll_local_import.setOnClickListener(new View.OnClickListener() {
+
             @Override
-            public void onClick(View v) {
+            public void doLocalImport() {
                 if (StorageUtils.isStoragePermissionGranted(context,activity)) {
                     showFileSelector();
-                    manageOptions.dismiss();
                 }
             }
-        });
-        ll_delete_select.setOnClickListener(new View.OnClickListener() {
+
             @Override
-            public void onClick(View v) {
-                if (!adapter.isDelete_mode()){
-                    adapter.setDelete_mode(true);
-                    delete_select_title.setText("退出删除");
-                }else {
-                    adapter.setDelete_mode(false);
-                    delete_select_title.setText("删除模式");
+            public void doQRImport() {
+                if(CameraManager.checkPermission(activity,context)) {
+                    Intent intent = new Intent(NovelSourceManageActivity.this, CaptureActivity.class);
+                    startActivityForResult(intent, Constant.REQ_QR_CODE);
+                }
+            }
+
+            @Override
+            public void onDeleteModeChange(boolean isDeleteMode) {
+                adapter.setDelete_mode(isDeleteMode);
+            }
+
+            @Override
+            public void doCheckSource() {
+                BasicCounterHandler<Long> counterHandler = new BasicCounterHandler<>(
+                        new BasicCounterHandler.BasicCounterHandlerListener<Long>() {
+                    @Override
+                    public void onSuccess(Long result) {
+
+                    }
+
+                    @Override
+                    public void onError() {
+
+                    }
+
+                    @Override
+                    public void onAllProcessDone(ArrayList<Long> results) {
+                        int disable_num = 0;
+                        for (Long time : results) {
+                            if (time >= 99999)disable_num++;
+                        }
+                        String info = "书源校验完成";
+                        if (disable_num!=0)info += String.format(Locale.CHINA,",%d个书源失效",disable_num);
+                        Toast.makeText(context, info, Toast.LENGTH_SHORT).show();
+                        waitDialog.dismiss();
+                    }
+                });
+                counterHandler.setTotal_count(novelRequireList.size(),true);
+                waitDialog.setTitle("书源校验中");
+                waitDialog.show();
+                for (NovelRequire rule : novelRequireList) {
+                    CheckRespondThread check = new CheckRespondThread(rule,sourceDBTools);
+                    check.setHandler(counterHandler);
+                    check.start();
                 }
             }
         });
+        source_manage_menu.show(im_manage);
     }
 
     private void showEditTextForURL() {
@@ -256,7 +364,7 @@ public class NovelSourceManageActivity extends AppCompatActivity {
                 InputMethodManager im = (InputMethodManager) context
                         .getSystemService(Context.INPUT_METHOD_SERVICE);
                 im.toggleSoftInput(0, InputMethodManager.HIDE_NOT_ALWAYS);
-                manageOptions.dismiss();
+                source_manage_menu.dismiss();
             }
         });
         ImageButton imb_commit = popWiw.getContentView().findViewById(R.id.ITIM_link_commit);
@@ -272,6 +380,7 @@ public class NovelSourceManageActivity extends AppCompatActivity {
                     sourceGetterThread.setHandler(handler);
                     sourceGetterThread.start();
                     popWiw.dismiss();
+                    waitDialog.setTitle("书源载入中");
                     waitDialog.show();
                 }
             }
